requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==2.1.1
fastapi==0.100.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.7
pydantic==2.16.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
PyJWT==2.8.0
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.10.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
zhipuai==2.0.1
